+++ telegrambot.py
@@ -21,8 +21,8 @@ def enviar_arquivo(token, chat_id, caminho):
         with open(caminho, "rb") as f:
             await bot.send_document(chat_id=chat_id, document=f)
 
-    asyncio.run(enviar_xmls())
-    #metodos.log_mensagem("Ativar envio do arquivo")
+    #asyncio.run(enviar_xmls())
+    metodos.log_mensagem("Ativar envio do arquivo")
 
 def enviar_mensagem(token, chat_id, mensagem):
     # Criar instância do bot